refactor(chat-settings): extract shared preference sections
